requirements.txt
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
distro==1.9.0
Flask==3.1.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
openai==2.13.0
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.4
